InternetSpeedTestApp/requirements.txt
speedtest-cli
speedtest-cli
requests
